local/requirements.txt
opencv-python==4.9.0.80  
numpy==1.24.3
pillow==10.2.0
insightface==0.7.3
onnxruntime==1.16.3
